idea:
Using a new int array and fill all of that as -1.
Create a stack to save the index from given array.
Use a for loop to push the index into stack.
Use a while loop to check the stack's top element compare with current element
return the array as result.