Idea:
Use a integer array and List<List<Integer>> to save the sum and sequences.
Use two for loop, the outer for loop will use for current index.
The second for loop will go through from 0 to current index.
if the element is lower the current element, use the largest one with current.